Proposal screen changes

+++ app/src/main/java/com/blockchain/store/dao/database/model/Rules.java
@@ -8,6 +8,6 @@ public class Rules {
     @PrimaryKey
     public int id;
     public long minimumQuorum;
-    public int debatingPeriodDuration;
+    public long debatingPeriodDuration;
     public long requisiteMajority;
 }
